Fix: Support decimal values for full_baths column
- Change full_baths from INTEGER to NUMERIC(4,2) to support decimal values (2.5, 3.5, etc.)
- Fix import routes to use parseFloat() instead of parseInt() for full_baths
- Add formatBaths() helper function to properly display decimal bathroom counts
- Update display components (ApolloContactCard, LeadDetailModal) to format decimals correctly
- Drop and recreate dependent views (listings_unified, prospect_enrich_view) during migration
- Add migration files with proper view handling

Fixes issue where 2.5 baths was being stored/displayed as 25, 3.5 as 35, etc.

+++ LeadMap-main/app/api/admin/upload-csv/route.ts
@@ -72,7 +72,7 @@ export async function POST(request: NextRequest) {
         state: record.state || null,
         zip_code: record.zip_code || record.zip || null,
         beds: record.beds ? parseInt(record.beds) : null,
-        full_baths: record.full_baths ? parseInt(record.full_baths) : null,
+        full_baths: record.full_baths ? parseFloat(record.full_baths) : null,
         half_baths: record.half_baths ? parseInt(record.half_baths) : null,
         sqft: record.sqft ? parseInt(record.sqft) : null,
         year_built: record.year_built ? parseInt(record.year_built) : null,
